[CP] Introduce flex_cp_forward custom op for FlexAttention CP
The custom op will fetch the required K and V. Currently, the forward pass is just an all-gather, and the backward pass is a reduce-scatter.  While the logic is the same as all_gather_tensor_autograd, the custom op avoids the Autograd warning that wait_tensor() is registered to autograd.

For the next step, we should explore how to interpolate the required communication based on the information from BlockMask.


ghstack-source-id: d5259e166f933fc06b80b1fa51b84a9d69306812
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/163185